clean up context usage
