+++ src/components/screens/History/index.tsx
@@ -5,11 +5,11 @@ import type { IHistoryEntry } from '@model'
 import type { THistoryPageProps } from '@model/nav'
 import TopNav from '@nav/TopNav'
 import { FlashList } from '@shopify/flash-list'
-import { FocusClaimCtx } from '@src/context/FocusClaim'
+import { useFocusClaimContext } from '@src/context/FocusClaim'
 import { useThemeContext } from '@src/context/Theme'
 import { getHistory } from '@store/HistoryStore'
 import { globals } from '@styles'
-import { useContext, useEffect, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { StyleSheet, Text, View } from 'react-native'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
@@ -20,7 +20,7 @@ export default function HistoryPage({ navigation, route }: THistoryPageProps) {
 	const insets = useSafeAreaInsets()
 	const { t } = useTranslation(['common'])
 	const { color } = useThemeContext()
-	const { claimed } = useContext(FocusClaimCtx)
+	const { claimed } = useFocusClaimContext()
 	const [data, setData] = useState<Record<string, IHistoryEntry[]>>({})
 	// update history after claiming from clipboard when the app comes to the foreground
 	useEffect(() => {
